feat: 화면 캡처 로직 수정
