+++ src/components/face-challenge/facetime/MyParticipantTile.tsx
@@ -7,7 +7,7 @@ import {
 import { useEffect, useState } from 'react';
 import useFaceFilterWithModel from '../../../hooks/useFaceFilterWithModel';
 import { useFilterTypeStore } from '../../../store/filterTypeStore';
-import ScreenCapturer from '../utils/ScreenCapturer2';
+import ScreenCapturer from '../utils/ScreenCapturer';
 import ScreenRecorder from '../utils/ScreenRecorder';
 
 const encoder = new TextEncoder();
@@ -52,10 +52,9 @@ const MyParticipantTile = () => {
       <ScreenRecorder
         customMediaStream={trackRef.publication?.track?.mediaStream || null}
       />
-
       <ParticipantTile />
       <canvas className="absolute left-0 top-0" ref={canvasRef} />
-      <ScreenCapturer videoElement={videoElement} />
+      <ScreenCapturer videoElement={videoElement} position={position} />
     </div>
   );
 };
